Alterações

+++ AlgoritmoGenetico/Models/Individuo.cs
@@ -6,7 +6,7 @@
         public Periodo Periodo_3_4 { get; set; }
         public Periodo Periodo_5_6 { get; set; }
         public Periodo Periodo_7_8 { get; set; }
-        public double Aptidao { get; set; }
+        public int Aptidao { get; set; }
 
         public Individuo()
         {
